BUG FIX: Incorrect operation if load_sample = False
Prior to this fix, if we created the database with load_samples = False, the client code fails to process listing or post responses.  This was because the listing and post processors tried to use existing database entries as a pro forma for new entries.  This fix eliminates that behaviour.  This fix addresses issue #2