BUG FIX: Client mocking results in missed messages
Running client mocking with several messages results in missed messages.  This was due to flaky code that checks the destination of incoming messages vs. the current callsign.  The code now waits until the callsign is set, refreshing the status with each check, before processing incoming messages.  This change fixes Issue #3.