Remove loading of database samples as default
The default setting for the database setup was to load sample data.  Once you start to use the client in real scenarios, the sample data causes a lot of confusion and makes the client behave strangely.  For these reasons, I have changed load_samples variable to False by default.  There was an issue doing this as the code used the sample data as a proforma for new messages but this requirement was removed by an earlier change.